Correction, AGAIN
I've done this exercise in Python yesterday (March 03, 2023) and in VisuAlg (~December 12, 2022) and something was wrong because when I checked the code and compare to another student, my calculation wasn't right. So, again, I changed somethings to run this correctly. I like to redo this exercises in Python because I'm practicing more than me last year.